fix: img collection

+++ controllers/Themes.js
@@ -1,12 +1,20 @@
 import Themes from "../models/Themes.js";
 
+import jwt from "jsonwebtoken";
+import bcrypt from "bcryptjs";
+const chechIsAdmin = (token) => {
+  const decoded = jwt.decode(token);
+  return decoded.role === "Admin";
+};
 class ThemeController {
   async create(req, res) {
     try {
-      const { theme, themeRu } = req.body;
+      const { theme, themeRu, token } = req.body;
       const variant = await Themes.findOne({ theme });
       const variant2 = await Themes.findOne({ themeRu });
-
+      if (!chechIsAdmin(token)) {
+        return res.status(400).json({ message: "You are not admin" });
+      }
       if (variant || variant2) {
         return res.status(400).json({
           message: "Such theme is already exist ",
@@ -23,9 +31,11 @@ class ThemeController {
   }
   async delete(req, res) {
     try {
-      const { _id } = req.params;
+      const { _id, token } = req.params;
       const theme = Themes.findOne(_id);
-
+      if (!chechIsAdmin(token)) {
+        return res.status(400).json({ message: "You are not admin" });
+      }
       await theme.deleteOne();
       return res.json(newTheme);
     } catch (error) {}
